+++ source/TimeSeries/MessageReceiver.IntegrationTests/integrationtest.local.settings.sample.json
